dockers/consumer/requirements.txt
pika==1.3.1
psycopg2-binary
python-dotenv
